requirements.txt
streamlit
langchain
openai
google-generativeai
python-dotenv
pytesseract
pdf2image
Pillow
reportlab
PyPDF2
tiktoken
